Migrate ProductResource to the v5 split-class layout
Splits the 466-line ProductResource into:

  ProductResource/
  ├── Schemas/ProductForm.php  ← form components, status shouts, helpers
  └── Tables/ProductTable.php  ← columns, filters, actions

ProductResource.php drops from 466 to ~160 lines: metadata, page
registration, widgets, global search overrides, eloquent query, and
one-line form()/table() delegations. Static helpers on ProductForm /
ProductTable drop the redundant Form/Table suffix (getSkuComponent vs
getSkuFormComponent) — the class name already conveys context.

Cross-resource consumers updated:
  - BrandResource/Pages/ManageBrandProducts → ProductTable::getNameColumn(),
    ProductTable::getSkuColumn()
  - ProductResource/Pages/ListProducts → ProductForm::getBaseNameComponent(),
    ProductForm::getProductTypeComponent(), getSkuComponent(),
    getBasePriceComponent()

Hook target moves from ProductResource::class to ProductForm::class /
ProductTable::class with method names configureForm / configureTable.
The extendSkuValidation hook keeps its name — only the form/table-level
hooks are renamed by the spec.

The interface-aware model resolution that previously relied on
BaseResource::getModel() (via ResolvesModelContract) now resolves the
contract directly inside ProductForm::getBaseNameComponent() via the
container — the split class doesn't extend BaseResource.
