+++ packages/admin/src/Filament/Resources/ProductResource/Schemas/ProductForm.php
@@ -1,0 +1,206 @@
+<?php
+
+namespace Lunar\Admin\Filament\Resources\ProductResource\Schemas;
+
+use Awcodes\Shout\Components\Shout;
+use Filament\Forms\Components\Select;
+use Filament\Forms\Components\TextInput;
+use Filament\Schemas\Components\Component;
+use Filament\Schemas\Components\Section;
+use Filament\Schemas\Schema;
+use Illuminate\Database\Eloquent\Model;
+use Lunar\Admin\Support\Concerns\CallsHooks;
+use Lunar\Admin\Support\Forms\Components\Attributes;
+use Lunar\Admin\Support\Forms\Components\Tags as TagsComponent;
+use Lunar\Admin\Support\Forms\Components\TranslatedText as TranslatedTextInput;
+use Lunar\Core\FieldTypes\Text;
+use Lunar\Core\FieldTypes\TranslatedText;
+use Lunar\Core\Models\Attribute;
+use Lunar\Core\Models\Contracts\Product as ProductContract;
+use Lunar\Core\Models\Currency;
+use Lunar\Core\Models\CustomerGroup;
+use Lunar\Core\Models\ProductVariant;
+use Lunar\Core\Models\Tag;
+
+class ProductForm
+{
+    use CallsHooks;
+
+    public static function configure(Schema $schema): Schema
+    {
+        return self::callStaticLunarHook(
+            'configureForm',
+            $schema
+                ->components([
+                    ...static::getStatusShouts(),
+                    Section::make()->schema(static::getMainComponents()),
+                    static::getAttributeDataComponent(),
+                    static::getVariantAttributeDataComponent(),
+                ])
+                ->columns(1),
+        );
+    }
+
+    public static function getMainComponents(): array
+    {
+        return [
+            static::getBrandComponent(),
+            static::getProductTypeComponent(),
+            static::getTagsComponent(),
+        ];
+    }
+
+    public static function getStatusShouts(): array
+    {
+        return [
+            Shout::make('product-status')
+                ->content(__('lunarpanel::product.status.unpublished.content'))
+                ->type('info')
+                ->hidden(fn (Model $record) => static::isPublished($record)),
+            Shout::make('product-customer-groups')
+                ->content(__('lunarpanel::product.status.availability.customer_groups'))
+                ->type('warning')
+                ->hidden(fn (Model $record) => ! static::isPublished($record) || static::hasEnabledCustomerGroup($record)),
+            Shout::make('product-no-default-customer-group')
+                ->content(__('lunarpanel::product.status.availability.no_default_customer_group'))
+                ->type('warning')
+                ->hidden(fn (Model $record) => ! static::isPublished($record)
+                    || ! static::hasEnabledCustomerGroup($record)
+                    || (bool) CustomerGroup::modelClass()::getDefault()
+                ),
+            Shout::make('product-hidden-from-guests')
+                ->content(__('lunarpanel::product.status.availability.hidden_from_guests'))
+                ->type('warning')
+                ->hidden(fn (Model $record) => ! static::isPublished($record)
+                    || ! static::hasEnabledCustomerGroup($record)
+                    || ! CustomerGroup::modelClass()::getDefault()
+                    || static::isDefaultGroupVisibleToGuests($record)
+                ),
+            Shout::make('product-channels')
+                ->content(__('lunarpanel::product.status.availability.channels'))
+                ->type('warning')
+                ->hidden(fn (Model $record) => ! static::isPublished($record) || $record->channels()->where('enabled', true)->count()),
+        ];
+    }
+
+    public static function getSkuValidation(): array
+    {
+        return self::callStaticLunarHook('extendSkuValidation', [
+            'required' => true,
+            'unique' => true,
+        ]);
+    }
+
+    public static function getSkuComponent(): Component
+    {
+        $validation = static::getSkuValidation();
+
+        $input = TextInput::make('sku')
+            ->label(__('lunarpanel::product.form.sku.label'))
+            ->required($validation['required'] ?? false);
+
+        if ($validation['unique'] ?? false) {
+            $input->unique(fn () => (new ProductVariant)->getTable());
+        }
+
+        return $input;
+    }
+
+    public static function getBasePriceComponent(): Component
+    {
+        $currency = Currency::getDefault();
+
+        return TextInput::make('base_price')
+            ->numeric()
+            ->prefix($currency->code)
+            ->rules([
+                'min:'.(1 / $currency->factor),
+                "decimal:0,{$currency->decimal_places}",
+            ])
+            ->required();
+    }
+
+    public static function getBaseNameComponent(): Component
+    {
+        $model = app()->get(ProductContract::class)::class;
+
+        $nameType = Attribute::whereHandle('name')
+            ->whereAttributeType($model::morphName())
+            ->first()?->type ?: TranslatedText::class;
+
+        $component = TranslatedTextInput::make('name');
+
+        if ($nameType == Text::class) {
+            $component = TextInput::make('name');
+        }
+
+        return $component
+            ->label(__('lunarpanel::product.form.name.label'))
+            ->required();
+    }
+
+    public static function getBrandComponent(): Component
+    {
+        return Select::make('brand_id')
+            ->label(__('lunarpanel::product.form.brand.label'))
+            ->relationship('brand', 'name')
+            ->searchable()
+            ->preload()
+            ->createOptionForm([
+                TextInput::make('name')->required(),
+            ]);
+    }
+
+    public static function getProductTypeComponent(): Component
+    {
+        return Select::make('product_type_id')
+            ->label(__('lunarpanel::product.form.producttype.label'))
+            ->relationship('productType', 'name')
+            ->searchable()
+            ->preload()
+            ->live()
+            ->required();
+    }
+
+    public static function getTagsComponent(): Component
+    {
+        return TagsComponent::make('tags')
+            ->suggestions(Tag::all()->pluck('value')->all())
+            ->splitKeys(['Tab', ','])
+            ->label(__('lunarpanel::product.form.tags.label'))
+            ->helperText(__('lunarpanel::product.form.tags.helper_text'));
+    }
+
+    public static function getAttributeDataComponent(): Component
+    {
+        return Attributes::make();
+    }
+
+    public static function getVariantAttributeDataComponent(): Component
+    {
+        return Attributes::make()
+            ->using(ProductVariant::class)
+            ->relationship('variant')
+            ->hidden(fn (ProductContract $record) => $record->hasVariants);
+    }
+
+    protected static function isPublished(?Model $record): bool
+    {
+        return $record?->status === 'published';
+    }
+
+    protected static function hasEnabledCustomerGroup(Model $record): bool
+    {
+        return $record->customerGroups()->where('enabled', true)->exists();
+    }
+
+    protected static function isDefaultGroupVisibleToGuests(Model $record): bool
+    {
+        $default = CustomerGroup::modelClass()::getDefault();
+
+        return $default && $record->newQuery()
+            ->whereKey($record->getKey())
+            ->customerGroup($default)
+            ->exists();
+    }
+}
